+++ src/main/java/com/danrus/pas/config/ModConfig.java
@@ -1,6 +1,8 @@
 package com.danrus.pas.config;
 
+import com.danrus.pas.api.SkinData;
 import com.danrus.pas.config.categories.MainCategory;
+import com.danrus.pas.utils.VersioningUtils;
 import net.minecraft.client.gui.screens.Screen;
 
 import com.google.gson.GsonBuilder;
@@ -43,6 +45,9 @@ public class ModConfig {
     @SerialEntry
     public String defaultSkin = "";
 
+    @SerialEntry
+    public boolean showArmorStandWhileDownloading = true;
+
     public enum DOWNLOAD_STATUS_DISPLAY {
         NONE,
         ABOVE_HOTBAR,
@@ -69,6 +74,9 @@ public class ModConfig {
 
     public static void save() {
         HANDLER.save();
+        SkinData.DEFAULT_TEXTURE = HANDLER.instance().showArmorStandWhileDownloading
+            ? VersioningUtils.getResourceLocation("minecraft", "textures/entity/armorstand/wood.png")
+            : VersioningUtils.getResourceLocation("minecraft", "textures/entity/player/wide/steve.png");
     }
 
     public static void load() {